[FIRRTL/IMConstantProp] Make field sensitive
This commit changes IMConstantProp to handle passive aggregate types.
The main difference is to use `pair<Value, unsigned>` instead of `Value`
as keys of lattice values. As an implementation, `ValueAndLeafIndex` is used.
The second element represents the position of the value in **leafs** which
is different from firrtl field index because field index is the position at tree.